extend support for shapefile initiated in #4869 - added a GeoJsonMixin and moved most of GeoJsonCatalogItem code to it - added a ShapefileCatalogItem which handles loading of shapefiles - extended GeojsonCatalogItem to support loading of dropped zip file (currently not used) - added test for geojson and shapefile
