+++ lib/Models/registerCatalogMembers.ts
@@ -25,11 +25,13 @@ import GtfsCatalogItem from "./GtfsCatalogItem";
 import IonImageryCatalogItem from "./IonImageryCatalogItem";
 import KmlCatalogItem from "./KmlCatalogItem";
 import MagdaReference from "./MagdaReference";
+import MapboxVectorTileCatalogItem from "./MapboxVectorTileCatalogItem";
 import OpenStreetMapCatalogItem from "./OpenStreetMapCatalogItem";
 import SdmxJsonCatalogGroup from "./SdmxJson/SdmxJsonCatalogGroup";
 import SdmxJsonCatalogItem from "./SdmxJson/SdmxJsonCatalogItem";
 import SenapsLocationsCatalogItem from "./SenapsLocationsCatalogItem";
 import SensorObservationServiceCatalogItem from "./SensorObservationServiceCatalogItem";
+import ShapefileCatalogItem from "./ShapefileCatalogItem";
 import SplitItemReference from "./SplitItemReference";
 import StubCatalogItem from "./StubCatalogItem";
 import ThreddsCatalogGroup from "./ThreddsCatalogGroup";
@@ -38,7 +40,6 @@ import WebFeatureServiceCatalogGroup from "./WebFeatureServiceCatalogGroup";
 import WebFeatureServiceCatalogItem from "./WebFeatureServiceCatalogItem";
 import WebMapServiceCatalogGroup from "./WebMapServiceCatalogGroup";
 import WebMapServiceCatalogItem from "./WebMapServiceCatalogItem";
-import MapboxVectorTileCatalogItem from "./MapboxVectorTileCatalogItem";
 import WebMapTileServiceCatalogGroup from "./WebMapTileServiceCatalogGroup";
 import WebMapTileServiceCatalogItem from "./WebMapTileServiceCatalogItem";
 import WebProcessingServiceCatalogFunction from "./WebProcessingServiceCatalogFunction";
@@ -79,6 +80,10 @@ export default function registerCatalogMembers() {
   CatalogMemberFactory.register(GeoRssCatalogItem.type, GeoRssCatalogItem);
   CatalogMemberFactory.register(CsvCatalogItem.type, CsvCatalogItem);
   CatalogMemberFactory.register(CzmlCatalogItem.type, CzmlCatalogItem);
+  CatalogMemberFactory.register(
+    ShapefileCatalogItem.type,
+    ShapefileCatalogItem
+  );
   CatalogMemberFactory.register(ArcGisCatalogGroup.type, ArcGisCatalogGroup);
   CatalogMemberFactory.register(
     ArcGisMapServerCatalogItem.type,
@@ -213,7 +218,7 @@ export default function registerCatalogMembers() {
   // We try to convert zipped shapefiles to geojson
   UrlToCatalogMemberMapping.register(
     matchesExtension("zip"),
-    GeoJsonCatalogItem.type
+    ShapefileCatalogItem.type
   );
 
   // These items work by trying to match a URL, then loading the data. If it fails, they move on.
