+++ lib/Traits/GeoJsonTraits.ts
@@ -1,0 +1,31 @@
+import CatalogMemberTraits from "./CatalogMemberTraits";
+import FeatureInfoTraits from "./FeatureInfoTraits";
+import MappableTraits from "./MappableTraits";
+import mixTraits from "./mixTraits";
+import objectTrait from "./objectTrait";
+import primitiveTrait from "./primitiveTrait";
+import StyleTraits from "./StyleTraits";
+import UrlTraits from "./UrlTraits";
+
+export class GeoJsonTraits extends mixTraits(
+  FeatureInfoTraits,
+  UrlTraits,
+  MappableTraits,
+  CatalogMemberTraits
+) {
+  @objectTrait({
+    type: StyleTraits,
+    name: "Style",
+    description:
+      "Styling rules that follow [simplestyle-spec](https://github.com/mapbox/simplestyle-spec)"
+  })
+  style?: StyleTraits;
+
+  @primitiveTrait({
+    type: "boolean",
+    name: "Clamp to Ground",
+    description:
+      "Whether the features in this GeoJSON should be clamped to the terrain surface."
+  })
+  clampToGround: boolean = true;
+}
