+++ lib/Traits/ShapefileCatalogItemTraits.ts
@@ -1,0 +1,6 @@
+import { GeoJsonTraits } from "./GeoJsonTraits";
+import mixTraits from "./mixTraits";
+
+export default class ShapefileCatalogItemTraits extends mixTraits(
+  GeoJsonTraits
+) {}
